fix: PR observations to break apart into smaller components
